fix the decimal transfer

+++ src/main/java/com/webank/webase/front/trade/request/SendReq.java
@@ -2,6 +2,7 @@ package com.webank.webase.front.trade.request;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 
 @Data
@@ -9,7 +10,7 @@ public class SendReq {
 
     private String from;
     private String to;
-    private  BigInteger value;
+    private BigDecimal value;
     private BigInteger minUnit;
     private String data;
     private BigInteger assetId;
